requirements.txt
gunicorn==20.1.0
starlette==0.13.6
uvicorn[standard]==0.13.4
opentelemetry-distro[otlp]==0.23b2
opentelemetry-instrumentation-starlette==0.23b2
